requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.0
billiard==3.6.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coala-utils==0.5.1
cycler==0.10.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
Pillow==6.1.0
pycodestyle==2.5.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
soupsieve==1.9.3
SQLAlchemy==1.3.8
textblob==0.15.3
twitterscraper==1.3.1
urllib3==1.25.3
Werkzeug==0.15.5
wordcloud==1.5.0
